dev: [General] Add validtaion for contacts and user routes

+++ mycontacts-backend/controllers/contactController.js
@@ -3,15 +3,15 @@ const Contact = require("../models/contactModel");
 
 //@desc Get all contacts
 //@route GET /api/contacts
-//@access public
+//@access private
 const getContacts = asyncHandler(async (req, res) => {
-    const contacts = await Contact.find();
+    const contacts = await Contact.find({user_id: req.user.id});
     res.status(200).json(contacts);
 });
 
 //@desc Create new contact
 //@route POST /api/contacts
-//@access public
+//@access private
 const createContact = asyncHandler(async (req, res) => {
     console.log("The request body is: ", req.body);
     const {name, email, phone} = req.body;
@@ -23,13 +23,14 @@ const createContact = asyncHandler(async (req, res) => {
         name,
         email,
         phone,
+        user_id: req.user.id
     });
     res.status(201).json(contact);
 });
 
 //@desc Get contact
 //@route GET /api/contacts/:id
-//@access public
+//@access private
 const getContact = asyncHandler(async (req, res) => {
     const contact = await Contact.findById(req.params.id);
     if (!contact) {
@@ -40,13 +41,18 @@ const getContact = asyncHandler(async (req, res) => {
 });
 
 //@desc Update contact
-//@route POST /api/contacts/:id
-//@access public
+//@route PUT /api/contacts/:id
+//@access private
 const updateContact = asyncHandler(async (req, res) => {
     const contact = await Contact.findById(req.params.id);
     if (!contact) {
         res.status(404);
         throw new Error("Contact not found");
+    }
+
+    if (contact.user_id.toString() !== req.user.id) {
+        res.status(403);
+        throw new Error("User don't have permission to update other user contacts");
     }
 
     const updatedContact = await Contact.findByIdAndUpdate(
@@ -60,12 +66,17 @@ const updateContact = asyncHandler(async (req, res) => {
 
 //@desc Delete new contact
 //@route POST /api/contacts/:id
-//@access public
+//@access private
 const deleteContact = asyncHandler(async (req, res) => {
     const contact = await Contact.findById(req.params.id);
     if (!contact) {
         res.status(404);
         throw new Error("Contact not found");
+    }
+
+    if (contact.user_id.toString() !== req.user.id) {
+        res.status(403);
+        throw new Error("User don't have permission to update other user contacts");
     }
 
     await Contact.deleteOne({ _id: req.params.id });
